requirements.txt
#This file is a list of all pip requirements for your skill (if needed)
gensim
soundcloud
python-vlc
